Fixed Customizer, Added Editor CSS, Added a Test Project
Customizer previously did not correctly update images to their default values stored in the content. This is now fixed.

Implemented Editor Styles. The view of the editor can now be styled to look similar to the final site.

Added Templatetest1 to test some changes to the template independent of the main Moo Copy project